field & message + fixed float for coordonate

+++ backend/src/schema/field.py
@@ -8,7 +8,8 @@ class FieldCreate(BaseModel):
     id_user: UUID
     name: str
     description: str
-    location: List[int]
+    location: List[float]
+    # data: set
     img_path: str
